TINY-4612: Clean up inconsistent code location
